refactor: remove node email feature from project. Migrated to a email micro service itself
