+++ package.json
@@ -16,19 +16,19 @@
     "start": "yarn build && node ./build/index.js",
     "pm2": "pm2 start processes.json --no-daemon",
     "start:docker": "docker-compose up",
-    "start:docker:build": "docker-compose up --build"
+    "start:docker:build": "docker-compose up --build",
+    "lint": "eslint ./src"
   },
   "engines": {
     "node": ">=12.16.1"
   },
   "dependencies": {
+    "axios": "^0.20.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "ejs": "^3.1.5",
-    "email-templates": "^7.1.1",
     "express": "^4.17.1",
-    "nodemailer": "^6.4.11",
     "pg": "^8.3.3",
     "pg-hstore": "^2.3.3",
     "pm2": "^4.5.0",
