--- src/email-sender/templates/purchase/style.css
@@ -1,130 +1,0 @@
-.email-container {
-  width: 100%;
-  background-color: #f8f5f2;
-  font-size: 14px;
-  line-height: 30px;
-  font-family: 'Roboto', sans-serif;
-  text-align: center;
-}
-
-.email-wrapper {
-  width: 570px;
-  max-width: 570px;
-  margin: auto;
-  box-sizing: border-box;
-  padding-top: 50px;
-}
-
-.header-container {
-  width: 100%;
-  display: flex;
-  background-color: #ff8000;
-  height: 120px;
-}
-
-.header-link {
-  margin: auto;
-}
-
-.header-text {
-  color: white;
-  margin: auto;
-  margin-left: 10px;
-  font-size: 25px;
-  font-weight: bold;
-}
-
-.title {
-  text-align: center;
-  margin-top: 0;
-  margin-bottom: 45px;
-  padding-top: 35px;
-}
-
-.link {
-  text-decoration: none;
-}
-
-.page-link {
-  color: #ff8000;
-}
-
-.emai-body {
-  padding: 0px 5%;
-  background-color: white;
-}
-
-.divider {
-  border-bottom: 1px solid #e9e9e9;
-  line-height: 20px;
-  width: 100%;
-}
-.purchase-list {
-  width: 100%;
-  background-color: #f7f5f1;
-}
-.logo-img {
-  width: 100%;
-  margin: auto;
-  background: linear-gradient(to top, #f7f5f0 50%, white 50%);
-}
-.img {
-  border-radius: 50%;
-  padding: 16px;
-  border: 1px solid #ff8000;
-  background: white;
-}
-.list-item {
-  display: flex;
-  padding: 0 5%;
-}
-
-.list-item-quantity {
-  width: 20%;
-  text-align: left;
-}
-.list-item-name {
-  width: 50%;
-}
-.list-item-total-amount {
-  width: 30%;
-  text-align: right;
-}
-.purchase-title {
-  text-align: left;
-  padding: 0 5% 15px;
-  font-size: 1.5em;
-  font-weight: 500;
-}
-.info-item {
-  display: flex;
-  padding: 0 5%;
-}
-.info-text {
-  width: 50%;
-  text-align: left;
-}
-.info-value {
-  width: 50%;
-  text-align: right;
-}
-
-.order-title {
-  text-align: left;
-  padding-top: 15px;
-  font-size: 1.5em;
-  font-weight: 500;
-}
-
-.order-details {
-  background-color: white;
-  border: 1px solid #e9e9e9;
-  margin-top: 30px;
-  margin-bottom: 50px;
-  padding: 0 5%;
-}
-
-.order-details-container {
-  text-align: left;
-  margin-bottom: 20px;
-}
